refactor(drama): partition src into logic/ and effects/ domains
Mechanical, behavior-preserving re-home of the former `lib/` grab-bag: pure
data and helpers (constants, rnd/starsFor/likelihoodLabel/clamp) move to
`logic/`; everything that touches `document`/`browser`/`chrome`
(messaging, content-scraper, the popup FSM, keybindings, particles) moves
to `effects/`. `utils.ts`'s `el()` DOM-element factory is split out into
`effects/content/dom.ts` since it called `document.createElement` despite
living next to pure helpers under a "no DOM" file banner.

Update every `@drama/lib/*` import across components, the content/popup
entrypoints, and stories to the new paths. Add a README to each domain
dir stating the browser/DOM import contract, and update the extension
README's Project Structure section to reflect the split.

No runtime behavior change: build output is unchanged (the same
cross-entry shared chunk that existed before as `utils.js` now emits as
`dom.js`, for the same reason — `el()` is used by both the content-script
and popup bundles).

Closes #429
