+++ extensions/some-drama/src/components/form-opinionated/moment-tags/index.ts
@@ -2,7 +2,7 @@
 // Owns: chip-based tag selector. Converts the old checkbox grid into chips
 // with selection state + shimmer feedback on activation.
 
-import { el } from "@drama/lib/content/utils"
+import { el } from "@drama/effects/content/dom"
 import type { MomentTag } from "@drama/types"
 
 import type { JournalDraft } from "../use-drama-journal-state"
