+++ extensions/some-drama/src/components/form-opinionated/preview-card/index.ts
@@ -4,7 +4,7 @@
 // from the current JournalDraft.
 
 import type { JournalDraft } from "@drama/components/form-opinionated/use-drama-journal-state"
-import { el } from "@drama/lib/content/utils"
+import { el } from "@drama/effects/content/dom"
 
 const TAG_LABELS: Record<string, string> = {
   confession: "Confession",
