+++ extensions/some-drama/src/components/form-opinionated/transition-editor/index.ts
@@ -3,7 +3,7 @@
 // (DramaEntryOpinionated.transition is singular — matches popup-renderer's
 // getTransition() contract). No querySelectorAll, no row-array.
 
-import { el } from "@drama/lib/content/utils"
+import { el } from "@drama/effects/content/dom"
 
 import type { JournalDraft } from "../use-drama-journal-state"
