+++ extensions/some-drama/src/components/slideshow/index.ts
@@ -11,11 +11,12 @@
 //   applyAtmosphereState(), applyRatingState(), applySummaryState(),
 //   setBubbleVisible(), setBubbleQuote().
 
+import { el } from "@drama/effects/content/dom"
 import {
   FALLBACK_QUOTES,
   SLIDE_INTERVAL_MS,
-} from "@drama/lib/content/constants"
-import { el, starsFor } from "@drama/lib/content/utils"
+} from "@drama/logic/content/constants"
+import { starsFor } from "@drama/logic/content/utils"
 import type { CardState, MomentTag } from "@drama/types"
 
 // Display glyphs for tags — cosmetic only, mirrors form-opinionated's TAG_META.
