+++ extensions/some-drama/src/effects/content/keybindings.ts
