+++ extensions/some-drama/src/effects/popup/content-scraper.ts
